pyproject.toml
[project]
name = "pdfreducer"
version = "0.1.0"
description = "A PDF size reduction tool with CLI and web interface"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pikepdf>=9.0.0",
    "Pillow>=10.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[project.scripts]
pdfreducer = "pdfreducer.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["pdfreducer*"]

[tool.setuptools.package-data]
"pdfreducer.web" = ["static/*"]
